cmd/wit-bindgen-go: rename describe command to wit
